Add TB for LDC Setup/Test
