+++ siepiclab/drivers/__init__.py
@@ -11,3 +11,4 @@ from PolCtrl_keysight import PolCtrl_keysight
 from PowerMonitor_keysight import PowerMonitor_keysight
 from fls_keysight import fls_keysight
 from tls_keysight import tls_keysight
+from ldc_srs_ldc500 import ldc_srs_ldc500
